Android: Always capitalise first letter, and fix "Timeline" being shown as "go5" on release builds
partial other impls, and fix web build too

+++ shared/src/jvmMain/java/org/mlm/mages/matrix/MatrixPort.jvm.kt
@@ -1649,7 +1649,7 @@ class RustMatrixPort : MatrixPort, VerificationService {
         withClient {
             runCatching {
                 val overview = it.mediaCacheOverview()
-                MediaCacheOverview(totalBytes = overview.totalBytes.toULong())
+                MediaCacheOverview(totalBytes = overview.totalBytes)
             }.getOrNull()
         }
     }
